Move hooks to web3-react
