+++ src/cow-react/modules/account/containers/AccountDetails/index.tsx
@@ -56,7 +56,7 @@ import {
 import { isMobile } from 'utils/userAgent'
 import UnsupporthedNetworkMessage from 'components/UnsupportedNetworkMessage'
 import { SupportedChainId as ChainId } from 'constants/chains'
-import { useDisconnectWallet } from '@cow/modules/wallet/api/hooks/useDisconnectWallet'
+import { useDisconnectWallet } from '@cow/modules/wallet/web3-react/hooks/useDisconnectWallet'
 
 export const NETWORK_LABELS: { [chainId in ChainId]?: string } = {
   // [ChainId.RINKEBY]: 'Rinkeby',
