--- src/cow-react/modules/wallet/api/index.ts
@@ -1,3 +1,0 @@
-export * from '../web3-react/Web3Status'
-export * from '../web3-react/containers/WalletModal'
-export * from './hooks/useWalletInfo'
