+++ src/cow-react/modules/wallet/index.ts
@@ -1,1 +1,8 @@
-export * from './api'
+
+export * from './web3-react/hooks/useWalletInfo'
+export * from './web3-react/hooks/useDisconnectWallet'
+
+export * from './web3-react/Web3Status'
+export * from './web3-react/containers/WalletModal'
+export * from './web3-react/containers/Identicon'
+export * from './web3-react/containers/AddToMetamask'
